Redid the clock and moved some stuff
