--- Project-Verruckt/Assets/Clock/3dobjects/Materials.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 2e5c48e927057c54ea7c8daad5782d84
-folderAsset: yes
-timeCreated: 1520796020
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
