--- Project-Verruckt/Assets/Clock/3dobjects/Materials/clock_main.mat.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 099241b3f9881c54d854dcd4fcbda0d9
-NativeFormatImporter:
-  mainObjectFileID: -1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
