--- Project-Verruckt/Assets/Clock/3dobjects/Materials/wanduhr-uhr glass.mat.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 14a00a7937103ca43875d11ae5e6f620
-timeCreated: 1520796023
-licenseType: Store
-NativeFormatImporter:
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
