--- Project-Verruckt/Assets/Clock/Prefabs.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 4c96bd6e48524654b98083950270b532
-folderAsset: yes
-timeCreated: 1520796020
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
