--- Project-Verruckt/Assets/Clock/Prefabs/clock.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 595cf544943e29a4d9a7b813e4218a3c
-NativeFormatImporter:
-  mainObjectFileID: -1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
