--- Project-Verruckt/Assets/Clock/Scripts.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 4e19939fee9aa3c4e8d153dda9741e96
-folderAsset: yes
-timeCreated: 1520796020
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
