--- Project-Verruckt/Assets/Clock/Scripts/Clock.cs.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 3a0db685f114bbf4e837ee81aae56661
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
