--- Project-Verruckt/Assets/Clock/example_scene.unity.meta
@@ -1,6 +1,0 @@
-fileFormatVersion: 2
-guid: 9d1dfbf3196052d4fba6e88539420f4e
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
