+++ Project-Verruckt/Assets/Downloaded Assets/Clock.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 6699722f636c63f41b358614ed14dd92
+folderAsset: yes
+timeCreated: 1520796020
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
