+++ Project-Verruckt/Assets/Downloaded Assets/Clock/3dobjects.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 4c3d13a31775aaa4082bceda011da650
+folderAsset: yes
+timeCreated: 1520796020
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
