+++ Project-Verruckt/Assets/Downloaded Assets/Clock/3dobjects/Materials/clock_glass.mat.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 40a8ae58dd3f1d5499694777e3945a80
+NativeFormatImporter:
+  mainObjectFileID: -1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
