+++ Project-Verruckt/Assets/Downloaded Assets/Clock/3dobjects/Materials/wanduhr-uhr.mat.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 50d35e73fd7c0a54e8709d22c0dd3a6a
+timeCreated: 1520796023
+licenseType: Store
+NativeFormatImporter:
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
