+++ Project-Verruckt/Assets/Downloaded Assets/Clock/Scripts/Clock.cs.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 3a0db685f114bbf4e837ee81aae56661
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
