+++ Project-Verruckt/Assets/Downloaded Assets/Clock/Textures.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9dfac55a45958304eb7a3e6b1f3b5074
+folderAsset: yes
+timeCreated: 1520796020
+licenseType: Store
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
