+++ Project-Verruckt/Assets/Downloaded Assets/Clock/example_scene.unity.meta
@@ -1,0 +1,6 @@
+fileFormatVersion: 2
+guid: 9d1dfbf3196052d4fba6e88539420f4e
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
