+++ Project-Verruckt/Assets/Downloaded Assets/Clock/example_sceneSettings.lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9356de8aed8406644b0d3321476fec59
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4890085278179872738
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
