+++ Project-Verruckt/Assets/Music and Sounds/love 1.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: bdf9ef88b4c7d3545bdbec42f4d78bd2
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
